SOFT: Fixed unregister distance message handler bug in OneStepEvent. Finished distance based OneStepEvent experiment.

+++ src/experiments/experiment.h
@@ -165,7 +165,7 @@ class Experiment
     long m_ZeroDistance;													/**< Zero distance */
     long m_StartLength;													/**< Distance where the experiment starts. */
     long m_CurrentDistance;											/**< Current distance of the stage frame */
-    long m_CurrentForce;												/**< Current force */
+    double m_CurrentForce;												/**< Current force */
     std::vector<long> m_CurrentPositions;				/**< Vector with the current stage positions */
     bool m_ExitFlag;														/**< Flag indicating that the experiment should stop imediatly */
     double m_Area;																													/**< Area size of the sample. */
